Move users into a database
