+++ heutagogy/auth.py
@@ -1,31 +1,11 @@
 from heutagogy import app
+from heutagogy.persistence import User
 from flask_jwt import JWT
 from werkzeug.security import safe_str_cmp
 
 
-class User(object):
-    def __init__(self, username, password):
-        self.id = username
-        self.username = username
-        self.password = password
-
-    def __str__(self):
-        return "User(id='%s')" % self.id
-
-    def __getitem__(self, key):
-        return {
-            'id': self.id,
-            'username': self.username,
-            'password': self.password
-        }.get(key, None)
-
-
 def get_user(username):
-    record = app.config['USERS'].get(username, None)
-    if record:
-        return User(username, record['password'])
-    else:
-        return None
+    return User.query.filter_by(username=username).first()
 
 
 def authenticate(username, password):
